+++ src/main/java/de/maxhenkel/corpse/gui/CorpseContainerFactory.java
@@ -7,8 +7,7 @@ import net.minecraft.world.entity.player.Inventory;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.Items;
 import net.minecraft.world.phys.AABB;
-import net.minecraftforge.network.IContainerFactory;
-
+import net.neoforged.neoforge.network.IContainerFactory;
 import java.util.Optional;
 import java.util.UUID;