+++ src/main/java/de/maxhenkel/corpse/gui/CorpseInventoryContainer.java
@@ -12,9 +12,8 @@ import net.minecraft.world.Container;
 import net.minecraft.world.entity.player.Inventory;
 import net.minecraft.world.inventory.InventoryMenu;
 import net.minecraft.world.item.ItemStack;
-import net.minecraftforge.api.distmarker.Dist;
-import net.minecraftforge.api.distmarker.OnlyIn;
-
+import net.neoforged.api.distmarker.Dist;
+import net.neoforged.api.distmarker.OnlyIn;
 import java.util.List;
 
 public class CorpseInventoryContainer extends CorpseContainerBase implements ITransferrable {