+++ src/main/java/de/maxhenkel/corpse/net/MessageSwitchInventoryPage.java
@@ -4,8 +4,8 @@ import de.maxhenkel.corelib.net.Message;
 import de.maxhenkel.corpse.gui.CorpseAdditionalContainer;
 import net.minecraft.network.FriendlyByteBuf;
 import net.minecraft.world.inventory.AbstractContainerMenu;
-import net.minecraftforge.api.distmarker.Dist;
-import net.minecraftforge.event.network.CustomPayloadEvent;
+import net.neoforged.api.distmarker.Dist;
+import net.neoforged.neoforge.network.NetworkEvent;
 
 public class MessageSwitchInventoryPage implements Message {
 
@@ -26,7 +26,7 @@ public class MessageSwitchInventoryPage implements Message {
     }
 
     @Override
-    public void executeServerSide(CustomPayloadEvent.Context context) {
+    public void executeServerSide(NetworkEvent.Context context) {
         AbstractContainerMenu container = context.getSender().containerMenu;
         if (container instanceof CorpseAdditionalContainer) {
             CorpseAdditionalContainer containerCorpse = (CorpseAdditionalContainer) container;